requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
python-dotenv==1.0.1
apscheduler==3.10.4
pytesseract==0.3.13
Pillow==10.4.0
python-multipart==0.0.9
pydantic==2.8.2
httpx==0.27.2
